arm64: dts: hi3660-hikey960: add kim and btwilink nodes for bluetooth obsolete
Method of kim and btwilink is still in use in v4.9 kernel, so copy from v4.4.
One caution in v4.9 kernel is: there are conflicts when both kim and uart4
request the same set of uart4 pins. In dmesg:

[    5.856338] pinctrl-single e896c000.pinmux: pin e896c0ec.0 already requested by fdf01000.serial; cannot claim for kim
[    5.866993] pinctrl-single e896c000.pinmux: pin-59 (kim) status -22
[    5.873275] pinctrl-single e896c000.pinmux: could not request pin 59 (e896c0ec.0) from group uart4_pmx_func  on device pinctrl-single
[    5.885305] kim kim: Error applying setting, reverse things back
[    5.891581] kim: probe of kim failed with error -22

So, when copying from v4.4, pinctrl properties are removed from kim node.

Note: However, in mainline kernel starting from v4.12, bluetooth are connected
through serdev. This is why this patch is labelled obsolete.

Signed-off-by: Guodong Xu <guodong.xu@linaro.org>