TEMP: dts: arm64: hi3660: disable spi2, xloader update pending
In HI3660, SPI2 belongs to secure world by default. Accessing spi2 in kernel
causes CPU mode error.

To fix that, need to grant access to normal world in xloader. An updated
xloader is now available at master of
  https://github.com/96boards-hikey/tools-images-hikey960

Need to submit it to AOSP as well. After submitted to AOSP, this commit
can be dropped.

Signed-off-by: Guodong Xu <guodong.xu@linaro.org>