dts: hikey960: Add UEFI reboot reason page to reserved memory
In an earlier commit, we enabled the reboot-reason that is used
with UEFI, however when using the HiSi bootloader, this results
in a warning on bootup as the driver tries to map a page that
is otherwise used normally by the kernel.

With UEFI, the memmap is modified to reserve the page, but with
the HiSi bootloader its just a standard page.

So add an entry to the reserved memory node to keep the kernel
from using the page even if we're using the HiSi bootloader.

Tested with UEFI and "reboot bootloader" still functions.
Tested with HiSi bootloader and no warning is seen.

Change-Id: Ieefbbdc3daecd80ffdcaf9e8a616699a3fce1a61
Signed-off-by: John Stultz <john.stultz@linaro.org>